decrypt is seperated , now easily you do other job, If you want keymatrix from main() function, It can be done... also decrypt you can do.. but here still , it do it simultaneous, but in GUI user may want separate encrypt facility and seperate decrypt facility

+++ mathUtilities.hpp
@@ -240,7 +240,7 @@ namespace Helper {
     {
         for (size_t i = 0; i < split; i++) {
             for (size_t j = 0; j < ONECOLUMN; j++) {
-                std::cout << std::setw(3) << pseudoVector[i][j] << " ";
+                std::cout << std::setw(3) << (int)pseudoVector[i][j] << " ";
             }
             std::cout << "  ";
         }
